requirements-gpu.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorful==0.5.8
contourpy==1.3.3
cryptography==44.0.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cycler==0.12.1
distlib==0.4.0
filelock==3.19.1
flwr==1.24.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.9.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-health-checking==1.76.0
idna==3.11
importlib_metadata==8.7.0
iterators==0.0.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.1
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
py-spy==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
ray==2.51.1
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
soundfile==0.13.1
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.3.0
tomli_w==1.2.0
torch==2.11.0.dev20260106+cu128
torchaudio==2.10.0.dev20260106+cu128
torchvision==0.25.0.dev20260106+cu128
tqdm==4.67.1
triton==3.6.0+git9844da95
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
virtualenv==20.35.4
vitaldb==1.5.8
wfdb==4.3.0
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
